Improve Application container + tests
The Application object now enforces its singleton-ness through it's constructor. This means the developer will be able to define their own boot sequence while still providing access to the instance statically. Additionally, the Loader is now resolved out of the container, allowing it to be swapped out and mocked in tests.