Append extension in class
This means the path passed in for the Assets class must not include an extension. We should add a method to make sure the string passed in doesn't already include an extension, and if it does, remove it (or throw an Exception). This makes the process of enqueuing minified scripts easier.