Initial Container implementation
The Container can have values and services set into it, and can
also define closures to return services or values. If the closures
aren't shared, they'll be executed every time the key is fetched
from the container. If they are, the value will be saved and
returned by future fetches.