Add initial Http implementation
This includes a router which functions as a wrapper over the WP-API, as
well as the helper classes to make it function and service provider to
register routes in. The developer would extend the RouterServiceProvider
and add all their routes, namespaces, etc. to the `routes` method. This
service provider would have to be registered with the application so it can
be called and have the router registered with the Application container.