Revamp tests for new Model functionality
The methods for testing the Models have been completely revamped. The
Post model has been deleted, with the stub models serving as examples
for developer implementation. The Model is now fully tested.

All of the key generating methods have been memoized for performance.
The method for determining whether a property is fillable has also
changed slightly, as we are not determining the custom post type via
method rather than property declaration. If something isn't fillable,
we now throw an Exception, to make it clear that you didn't succeed
in the assignment.